+++ lib/src/flutter_web_bluetooth_unsupported.dart
@@ -13,6 +13,8 @@ part 'bluetooth_characteristic_properties.dart';
 
 part 'bluetooth_default_uuids.dart';
 
+part 'bluetooth_descriptor.dart';
+
 part 'bluetooth_device.dart';
 
 part 'bluetooth_service.dart';